refactor(tests): move fine_or_create_activity tests into the correct file
